+++ src/components/loaders/ui/search.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { Skeleton } from "../../ui/skeleton";
+import { Skeleton } from "@/components/ui/skeleton";
 
 export default function SearchLoader(){
      return (